refactor(imports): update import paths and formatting for consistency
- Adjusted import paths in `index.ts`, `redmine.test.ts`, and `redmine.ts` to ensure they correctly reference the new structure of the `common` directory.
- Reformatted code in `Redmine.ts` and `utils.ts` for improved readability, including consistent spacing and line breaks.

These changes enhance code clarity and maintainability by standardizing import statements and improving overall formatting.
