+++ workers/main/src/common/utils.ts
@@ -1,8 +1,13 @@
-import { validationResult } from '../main/src/configs';
+import { validationResult } from '../configs';
 
-export const formatValidationIssues = (issues: { path: (string | number)[]; message: string }[]): string =>
+export const formatValidationIssues = (
+  issues: { path: (string | number)[]; message: string }[],
+): string =>
   issues
-    .map(({ path, message }) => `Missing or invalid environment variable: ${path.join('.') || '(unknown variable)'} (${message})`)
+    .map(
+      ({ path, message }) =>
+        `Missing or invalid environment variable: ${path.join('.') || '(unknown variable)'} (${message})`,
+    )
     .join('\n');
 
 export function validateEnv() {
